use primtive types only for both input and output

+++ src/prime/sieve/eratosthenes.rs
@@ -16,14 +16,14 @@
 /// let primes = eratosthenes_sequence(30);
 /// assert_eq!(primes, vec![2, 3, 5, 7, 11, 13, 17, 19, 23, 29]);
 /// ```
-pub fn eratosthenes_sequence(n: u128) -> Vec<u128> {
+pub fn eratosthenes_sequence(n: usize) -> Vec<u128> {
     let mut primes = vec![];
-    let mut sieve = vec![true; n as usize + 1];
-    for i in 2..n as usize + 1 {
+    let mut sieve = vec![true; n + 1];
+    for i in 2..n + 1 {
         if sieve[i] {
             primes.push(i as u128);
             let mut j = i * i;
-            while j < n as usize + 1 {
+            while j < n + 1 {
                 sieve[j] = false;
                 j += i;
             }
